+++ components/blog/BlogPostContainer/index.js
@@ -3,10 +3,8 @@ import TwoColumn from '@/components/containers/TwoColumn';
 import AuthorBio from '@/components/blog/AuthorBio';
 import RelatedOrLatestPosts from '@/components/blog/RelatedPosts';
 import PostContent from '@/components/blog/PostContent';
-import { useTheme } from 'styled-components';
 
 const BlogPostContainer = ({ post, relatedPosts, latestPosts }) => {
-  const theme = useTheme();
   const { user } = post;
   return (
     <>
@@ -26,10 +24,8 @@ const BlogPostContainer = ({ post, relatedPosts, latestPosts }) => {
         <TwoColumn
           title='Still got questions?'
           content='Feel free to contact us.'
-          image='/images/svg/square-brackets.svg'
-          altTag='Hashtag'
-          color={theme.colors.primaryContent}
-          bgColor={theme.colors.white}
+          image='/images/svg/square-brackets.svg' // Decorative Image
+          altTag='' // Intentionally left empty to comply with W3C accessibility standards.
           $contentType='questions'
           link='/contact'
           linkText='Contact us'