+++ CHANGELOG.md
@@ -163,5 +163,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - ButtonLink
   - SubmitButton
   - AuthorBio
-  
+  - BlogPostContainer
 - Updated ContactUsForm's checkbox wrapper from div to label to enhance its accessibility
+- Updated altTag from TwoColumns component in BlogPostContainer to comply with W3C standards